TGP-2890: mover formProviders and fix controller specs
