+++ app/controllers/categorisation/CategorisationPreparationController.scala
@@ -23,7 +23,7 @@ import models.Scenario.getResultAsInt
 import models.helper.CategorisationUpdate
 import models.ott.CategorisationInfo
 import models.{CategoryRecord, Mode, NormalMode, UserAnswers}
-import navigation.Navigator
+import navigation.CategorisationNavigator
 import org.apache.pekko.Done
 import pages.{CategorisationPreparationPage, HasSupplementaryUnitPage, RecategorisationPreparationPage}
 import play.api.i18n.MessagesApi
@@ -41,17 +41,17 @@ import scala.concurrent.{ExecutionContext, Future}
 import scala.util.{Success, Try}
 
 class CategorisationPreparationController @Inject() (
-  override val messagesApi: MessagesApi,
-  val controllerComponents: MessagesControllerComponents,
-  identify: IdentifierAction,
-  getData: DataRetrievalAction,
-  requireData: DataRequiredAction,
-  profileAuth: ProfileAuthenticateAction,
-  categorisationService: CategorisationService,
-  goodsRecordsConnector: GoodsRecordConnector,
-  sessionRepository: SessionRepository,
-  navigator: Navigator,
-  auditService: AuditService
+                                                      override val messagesApi: MessagesApi,
+                                                      val controllerComponents: MessagesControllerComponents,
+                                                      identify: IdentifierAction,
+                                                      getData: DataRetrievalAction,
+                                                      requireData: DataRequiredAction,
+                                                      profileAuth: ProfileAuthenticateAction,
+                                                      categorisationService: CategorisationService,
+                                                      goodsRecordsConnector: GoodsRecordConnector,
+                                                      sessionRepository: SessionRepository,
+                                                      navigator: CategorisationNavigator,
+                                                      auditService: AuditService
 )(implicit ec: ExecutionContext)
     extends BaseController {
