+++ app/forms/categorisation/HasSupplementaryUnitFormProvider.scala
@@ -14,22 +14,17 @@
  * limitations under the License.
  */
 
-package forms
+package forms.categorisation
 
 import forms.mappings.Mappings
-import javax.inject.Inject
 import play.api.data.Form
-import forms.mappings.helpers.FormatAnswers.removeWhitespace
-import models.StringFieldRegex
 
-class LongerCommodityCodeFormProvider @Inject() extends Mappings {
+import javax.inject.Inject
 
-  def apply(): Form[String] =
+class HasSupplementaryUnitFormProvider @Inject() extends Mappings {
+
+  def apply(): Form[Boolean] =
     Form(
-      "value" -> text("longerCommodityCode.error.required")
-        .transform(removeWhitespace, identity[String])
-        .verifying(
-          regexp(StringFieldRegex.commodityCodeAdditionalNumbersRegex, "longerCommodityCode.error.invalidFormat")
-        )
+      "value" -> boolean("hasSupplementaryUnit.error.required")
     )
 }
